add --version flag

+++ hubploy/helm.py
@@ -28,16 +28,18 @@ def helm_upgrade(
     namespace,
     chart,
     config_files,
-    config_overrides
+    config_overrides,
+    version
 ):
     # Clear charts and do a helm dep up before installing
     # Clearing charts is important so we don't deploy charts that
     # have been removed from requirements.yaml
     # FIXME: verify if this is actually true
-    shutil.rmtree(os.path.join(chart, 'charts'), ignore_errors=True)
-    subprocess.check_call([
-        'helm', 'dep', 'up'
-    ], cwd=chart)
+    if os.path.exists(chart):
+        shutil.rmtree(os.path.join(chart, 'charts'), ignore_errors=True)
+        subprocess.check_call([
+            'helm', 'dep', 'up'
+        ], cwd=chart)
 
     cmd = [
         'helm',
@@ -47,9 +49,11 @@ def helm_upgrade(
         '--install',
         '--namespace', namespace,
         name, chart,
-    ] 
-    cmd += itertools.chain(*[['-f', cf] for cf in config_files]) 
-    cmd += itertools.chain(*[['--set', f'{k}={v}'] for k, v in config_overrides.items()])
+    ]
+    if version:
+        cmd += ['--version', version]
+    cmd += itertools.chain(*[['-f', cf] for cf in config_files])
+    cmd += itertools.chain(*[['--set', v] for v in config_overrides])
     subprocess.check_call(cmd)
 
 
@@ -59,6 +63,7 @@ def deploy(
     environment,
     namespace=None,
     config_overrides=None,
+    version=None,
 ):
     """
     Deploy a JupyterHub.
@@ -106,7 +111,8 @@ def deploy(
         namespace,
         chart,
         config_files,
-        config_overrides
+        config_overrides,
+        version
     )
 
 
@@ -130,10 +136,13 @@ def main():
         '--set',
         action='append',
     )
+    argparser.add_argument(
+        '--version',
+    )
 
     args = argparser.parse_args()
 
-    deploy(args.deployment, args.chart, args.environment, args.namespace, args.set)
+    deploy(args.deployment, args.chart, args.environment, args.namespace, args.set, args.version)
 
 if __name__ == '__main__':
     main()
